+++ .eslintrc.js
@@ -1,3 +1,10 @@
 module.exports = {
-  extends: ["next"]
+  extends: ["next"],
+  plugins: ["react"],
+  rules: {
+    "react/forbid-dom-props": [
+      "warn",
+      { forbid: ["style"] }
+    ]
+  }
 };